Configuration III: the supertype model
The supertype model defines a model by which an XML document can be read
and converted into a tree which describes the layout of any other
configuration file syntactically. This allows for the future
implementation of configuration file structures which are governed by a
mod ResourceLocation file which describes the expected types,
properties, values, typecasting and structure of the configuration file
- as well as how to handle inconsistencies, missing properties, comments
and other semantic settings.